06 Trimestre 7/prueba/jjsystemPrueba/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.16.0
colorama==0.4.6
Django==5.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
iniconfig==2.0.0
mysqlclient==2.2.3
packaging==23.2
pluggy==1.4.0
pycparser==2.21
PyJWT==2.8.0
pytest==8.0.2
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
